Update: 修复了关于GetLong和 PopLong 的 bug

+++ ch04/rtda/local_vars.go
@@ -38,8 +38,8 @@ func (self LocalVars) SetLong(index uint, val int64) {
 }
 
 func (self LocalVars) GetLong(index uint) int64 {
-	low := self[index].num
-	high := self[index+1].num
+	low := uint32(self[index].num)
+	high := uint32(self[index+1].num)
 	return int64(high)<<32 | int64(low)
 }
